+++ packages/node-core/src/api.service.spec.ts
@@ -3,17 +3,18 @@
 
 import {EventEmitter2} from '@nestjs/event-emitter';
 import {delay} from '@subql/common';
-import {ProjectNetworkConfig} from '@subql/types-core';
+import {IEndpointConfig, ProjectNetworkConfig} from '@subql/types-core';
 import {ApiService, IApiConnectionSpecific} from './api.service';
 import {NodeConfig} from './configure';
 import {ConnectionPoolService, ConnectionPoolStateManager} from './indexer';
 
 class TestApiService extends ApiService {
   retryConnection(
-    createConnection: (endpoint: string) => Promise<IApiConnectionSpecific>,
+    createConnection: (endpoint: string, config: IEndpointConfig) => Promise<IApiConnectionSpecific>,
     network: ProjectNetworkConfig & {chainId: string},
     index: number,
     endpoint: string,
+    config: IEndpointConfig,
     postConnectedHook?: ((connection: IApiConnectionSpecific, endpoint: string, index: number) => void) | undefined
   ): void {
     /* No Op to avoid creating intervals/timeouts in tests*/
@@ -33,7 +34,7 @@ describe('ApiService', () => {
   it('should throw creating connections if all endpoints are invalid', async () => {
     await expect(
       apiService.createConnections(
-        {chainId: 'test', endpoint: ['fail', 'fail']} as any,
+        {chainId: 'test', endpoint: {fail: {}, fail2: {}}},
         (endpoint) => Promise.resolve({networkMeta: {chain: endpoint}} as any) // Fail meta validation
       )
     ).rejects.toThrow();
@@ -42,7 +43,7 @@ describe('ApiService', () => {
   it('should succeed creating connections if one endpoint is invalid', async () => {
     await expect(
       apiService.createConnections(
-        {chainId: 'test', endpoint: ['test', 'fail']} as any,
+        {chainId: 'test', endpoint: {test: {}, fail: {}}},
         (endpoint) => Promise.resolve({networkMeta: {chain: endpoint}} as any) // Fail meta validation
       )
     ).resolves.not.toThrow();
@@ -50,7 +51,7 @@ describe('ApiService', () => {
 
   it(`doesn't set network meta if a connection fails validation`, async () => {
     await expect(
-      apiService.createConnections({chainId: 'test', endpoint: ['fail']} as any, (endpoint) =>
+      apiService.createConnections({chainId: 'test', endpoint: ['fail']}, (endpoint) =>
         Promise.resolve({networkMeta: {chain: endpoint}} as any)
       )
     ).rejects.toThrow();
@@ -61,7 +62,7 @@ describe('ApiService', () => {
   it('should retry connections if they fail with something other than metadata', async () => {
     const retrySpy = jest.spyOn(apiService, 'retryConnection');
 
-    await apiService.createConnections({chainId: 'test', endpoint: ['test', 'fail']} as any, (endpoint) =>
+    await apiService.createConnections({chainId: 'test', endpoint: {test: {}, fail: {}}}, (endpoint) =>
       endpoint === 'test'
         ? Promise.resolve({networkMeta: {chain: endpoint}} as any) // At least one endpoint needs to succeed
         : Promise.reject(new Error('Test'))