+++ packages/node/src/indexer/project.service.spec.ts
@@ -24,7 +24,7 @@ function testSubqueryProject(): SubqueryProject {
     id: 'test',
     root: './',
     network: {
-      endpoint: ['https://polkadot.api.onfinality.io/public'],
+      endpoint: { 'https://polkadot.api.onfinality.io/public': {} },
       chainId:
         '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3',
     },