+++ packages/types-core/src/project/types.ts
@@ -12,7 +12,7 @@ import {BaseDataSource, BaseTemplateDataSource, ParentProject, RunnerSpecs} from
 export interface CommonSubqueryProject<
   N extends IProjectNetworkConfig = IProjectNetworkConfig,
   DS extends BaseDataSource = BaseDataSource,
-  T extends BaseTemplateDataSource<DS> = BaseTemplateDataSource<DS>
+  T extends BaseTemplateDataSource<DS> = BaseTemplateDataSource<DS>,
 > {
   /**
    * The repository of your SubQuery project.
@@ -91,7 +91,8 @@ export interface CommonSubqueryProject<
  * @interface
  * @extends {ProjectNetworkConfig}
  */
-export interface IProjectNetworkConfig extends ProjectNetworkConfig {
+export interface IProjectNetworkConfig<EndpointConfig extends IEndpointConfig = IEndpointConfig>
+  extends ProjectNetworkConfig<EndpointConfig> {
   /**
    * The unique identity of the chain.
    *
@@ -128,11 +129,17 @@ export interface IProjectManifest<D> {
   validate(): void;
 }
 
+/* Define specific behaviour for rate limits*/
+export interface IEndpointConfig {
+  /* Headers to be supplied with the requests to the endpoint */
+  headers?: Record<string, string>;
+}
+
 /**
  * Represents the network configuration for a project.
  * @interface
  */
-export interface ProjectNetworkConfig {
+export interface ProjectNetworkConfig<EndpointConfig extends IEndpointConfig = IEndpointConfig> {
   /**
    * The endpoint(s) for the network connection, which can be a single string or an array of strings.
    *
@@ -141,9 +148,9 @@ export interface ProjectNetworkConfig {
    * Public nodes may be rate limited, which can affect indexing speed
    * When developing your project we suggest adding a private API key
    *
-   * @type {string | string[]}
+   * @type {string | string[] | Record<string, IEndpointConfig> }
    */
-  endpoint: string | string[];
+  endpoint: string | string[] | Record<string, EndpointConfig>;
 
   /**
    * The SubQuery network dictionary endpoint (optional).