+++ packages/node-core/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- A more useful error message when failing to require modules from the VM (#2512)
+
+### Added
+- Support for endpoint configs (#2511)
 
 ## [13.0.2] - 2024-07-31
 ### Fixed